move some error models around
